Added shader background with basic colour control
Colour control is passed from GUI sliders to the fragment shader. There
was a problem that broke all the meshes, meaning you only got one slice
of the pie. This is because I’d used glPushMatrix instead of
ofPushMatrix. It’s still valid and compiled but I think that was
calling a particular function related to a different version of openGL
to that which the ofGLProgrammableRenderer uses, so it broke it. Fixing
these typos also fixed the error regarding closing the matrices. This
also fixed the problem with the GUI rotating for the first few seconds
after the app started.